+++ Enums/CropMode.cs
@@ -3,86 +3,86 @@
 		/// <summary>
 		/// South / Bottom
 		/// </summary>
-		public const string Bottom = "b";
+		public static readonly string Bottom = "b";
 
 		/// <summary>
 		/// South West / Bottom Left
 		/// </summary>
-		public const string BottomLeft = "bl";
+		public static readonly string BottomLeft = "bl";
 
 		/// <summary>
 		/// South East / Bottom Right
 		/// </summary>
-		public const string BottomRight = "br";
+		public static readonly string BottomRight = "br";
 
 		/// <summary>
 		/// Center - this is the default gravity or direction, and applied when the crop_mode parameter is left out, or an invalid value is passed.
 		/// </summary>
-		public const string Center = "c";
+		public static readonly string Center = "c";
 
 		/// <summary>
 		/// East / Right
 		/// </summary>
-		public const string East = "e";
+		public static readonly string East = "e";
 
 		/// <summary>
 		/// West / Left
 		/// </summary>
-		public const string Left = "l";
+		public static readonly string Left = "l";
 
 		/// <summary>
 		/// North / Top
 		/// </summary>
-		public const string North = "n";
+		public static readonly string North = "n";
 
 		/// <summary>
 		/// North East / Top Right
 		/// </summary>
-		public const string NorthEast = "ne";
+		public static readonly string NorthEast = "ne";
 
 		/// <summary>
 		/// North West / Top Left
 		/// </summary>
-		public const string NorthWest = "nw";
+		public static readonly string NorthWest = "nw";
 
 		/// <summary>
 		/// East / Right
 		/// </summary>
-		public const string Right = "r";
+		public static readonly string Right = "r";
 
 		/// <summary>
 		/// South / Bottom
 		/// </summary>
-		public const string South = "s";
+		public static readonly string South = "s";
 
 		/// <summary>
 		/// South East / Bottom Right
 		/// </summary>
-		public const string SouthEast = "se";
+		public static readonly string SouthEast = "se";
 
 		/// <summary>
 		/// South West / Bottom Left
 		/// </summary>
-		public const string SouthWest = "sw";
+		public static readonly string SouthWest = "sw";
 
 		/// <summary>
 		/// North / Top
 		/// </summary>
-		public const string Top = "t";
+		public static readonly string Top = "t";
 
 		/// <summary>
 		/// North West / Top Left
 		/// </summary>
-		public const string TopLeft = "tl";
+		public static readonly string TopLeft = "tl";
 
 		/// <summary>
 		/// North East / Top Right
 		/// </summary>
-		public const string TopRight = "tr";
+		public static readonly string TopRight = "tr";
 
 		/// <summary>
 		/// West / Left
 		/// </summary>
-		public const string West = "w";
+		public static readonly string West = "w";
 	}
 }