+++ Enums/Metadata.cs
@@ -9,7 +9,7 @@
 		/// - Xmp.dc.rights
 		/// - Xmp.photoshop.Credit
 		/// </summary>
-		public const string Copyright = "copyright";
+		public static readonly string Copyright = "copyright";
 
 		/// <summary>
 		/// Will preserve image creation date. The following tags will be preserved (if present):
@@ -18,7 +18,7 @@
 		/// - Exif.Photo.DateTimeOriginal
 		/// - Xmp.photoshop.DateCreated
 		/// </summary>
-		public const string Date = "date";
+		public static readonly string Date = "date";
 
 		/// <summary>
 		/// Will preserve location-specific information. The following tags will be preserved (if present):
@@ -29,18 +29,18 @@
 		/// - Exif.GPSInfo.GPSLongitudeRef
 		/// - Exif.GPSInfo.GPSVersionID
 		/// </summary>
-		public const string Geotag = "geotag";
+		public static readonly string Geotag = "geotag";
 
 		/// <summary>
 		/// Will preserve the orientation (rotation) mark. The following tag will be preserved (if present):
 		/// 
 		/// - Exif.Image.Orientation
 		/// </summary>
-		public const string Orientation = "orientation";
+		public static readonly string Orientation = "orientation";
 
 		/// <summary>
 		/// Will preserve the ICC colour profile. ICC colour profile information adds unnecessary bloat to images. However, preserving it can be necessary in extremely rare cases where removing this information could lead to a change in brightness and/or saturation of the resulting file.
 		/// </summary>
-		public const string Profile = "profile";
+		public static readonly string Profile = "profile";
 	}
 }